requirements.txt
# Microsoft Graph API
msal>=1.25.0
httpx>=0.25.0

# Data processing
pydantic>=2.0.0
python-dateutil>=2.8.2

# Integrations
jira>=3.5.0
atlassian-python-api>=3.41.0

# Email
secure-smtplib>=0.1.1

# Scheduling
apscheduler>=3.10.4

# CLI
click>=8.1.0
rich>=13.0.0

# Configuration
pyyaml>=6.0.1

# Async support
aiofiles>=23.0.0

# Existing dependencies (from mcp/requirements.txt)
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
